fix(app): dqa software keyboard round3
address software keyboard feedback

close RAUT-615
