+++ app/src/atoms/SoftwareKeyboard/CustomKeyboard/CustomKeyboard.stories.tsx
@@ -5,17 +5,18 @@ import {
   POSITION_ABSOLUTE,
   SPACING,
 } from '@opentrons/components'
-import { InputField } from '../InputField'
-import { Numpad } from './'
+import { InputField } from '../../InputField'
+import { CustomKeyboard } from '.'
+import '../../../styles.global.css'
 
 import type { Story, Meta } from '@storybook/react'
 
 export default {
-  title: 'Odd/Atoms/SoftwareKeyboard/Numpad',
-  component: Numpad,
+  title: 'ODD/Atoms/SoftwareKeyboard/CustomKeyboard',
+  component: CustomKeyboard,
 } as Meta
 
-const Template: Story<React.ComponentProps<typeof Numpad>> = args => {
+const Template: Story<React.ComponentProps<typeof CustomKeyboard>> = args => {
   const [showKeyboard, setShowKeyboard] = React.useState(false)
   const [value, setValue] = React.useState<string>('')
   const keyboardRef = React.useRef(null)
@@ -25,13 +26,13 @@ const Template: Story<React.ComponentProps<typeof Numpad>> = args => {
         <InputField
           value={value}
           type="text"
-          placeholder="When focusing, the numpad shows up"
+          placeholder="When focusing, the keyboard shows up"
           onFocus={() => setShowKeyboard(true)}
         />
       </form>
-      <Flex position={POSITION_ABSOLUTE} top="20%" width="15rem">
+      <Flex position={POSITION_ABSOLUTE} top="20%" width="55rem">
         {showKeyboard && (
-          <Numpad
+          <CustomKeyboard
             onChange={e => e != null && setValue(String(e))}
             keyboardRef={keyboardRef}
           />
@@ -41,4 +42,4 @@ const Template: Story<React.ComponentProps<typeof Numpad>> = args => {
   )
 }
 
-export const NormalSoftwareKeyboard = Template.bind({})
+export const CustomSoftwareKeyboard = Template.bind({})
