+++ app/src/atoms/SoftwareKeyboard/CustomKeyboard/__tests__/CustomKeyboard.test.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { fireEvent } from '@testing-library/react'
 import { renderHook } from '@testing-library/react-hooks'
 import { renderWithProviders } from '@opentrons/components'
-import { CustomKeyboard } from '..'
+import { CustomKeyboard } from '../'
 
 const render = (props: React.ComponentProps<typeof CustomKeyboard>) => {
   return renderWithProviders(<CustomKeyboard {...props} />)[0]
