+++ app/src/atoms/SoftwareKeyboard/NormalKeyboard/index.css
@@ -1,0 +1,5 @@
+/* button.hg-button.hg-button-backspace,
+button.hg-button.hg-button-abc,
+button.hg-button.hg-standardBtn {
+  color: red;
+} */
