+++ app/src/atoms/SoftwareKeyboard/NormalKeyboard/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import Keyboard from 'react-simple-keyboard'
-import { customDisplay } from './constants'
+import { customDisplay } from '../constants'
 
 interface NormalKeyboardProps {
   onChange: (input: string) => void
@@ -20,7 +20,7 @@ const customLayout = {
   shift: [
     'Q W E R T Y U I O P',
     '{numbers} A S D F G H J K L',
-    '{shift} Z X C V B N M {backspace}',
+    '{abc} Z X C V B N M {backspace}',
     '{space}',
   ],
   symbols: [
